Version Finale

+++ lib/ADDU.c
@@ -17,7 +17,7 @@ int exec(unsigned int* jump, pm_glob param, INST inst){
 	a = registre[inst.rs].content ; //On impose l'interprétation des valeurs des registres comme entiers codés sur 32 bits
 	b = registre[inst.rt].content ;	
 
-	registre[inst.rd].content = a + b ;
+	registre[inst.rd].content = (int32_t) a + b ;
 
 	return 0;
 }
